Replace hard-coded presets with system prompt presets

+++ backend/src/HEAL.HeuristicAgent.Web/Dtos/DataPointDto.cs
@@ -1,3 +1,3 @@
 ﻿namespace HEAL.HeuristicAgent.Web.Dtos;
 
-public sealed record DataPointDto(string Id, double Value);
+public sealed record DataPointDto(string Id, float Value);
